Split get_events into new file alongside get_updates
