+++ get_events.py
@@ -1,25 +1,16 @@
 import os
 import icalendar
 import requests
-import calendar
 from datetime import datetime, timezone
 from typing import List
 from icalendar import Event as IcalEvent
+from date_string import get_date_string
 
 CONTACT_EMAIL = os.environ.get("CONTACT_EMAIL")
 CHURCH_ID = os.environ.get("CHURCH_ID")
-CHURCH_LOGO_URL = os.environ.get("CHURCH_LOGO_URL")
-CHURCH_NAME = os.environ.get("CHURCH_NAME")
 
 if CHURCH_ID == None:
     raise Exception("Church ID is not defined")
-
-
-def get_date_string(date: datetime) -> str:
-    day = date.day
-    month = calendar.month_name[date.month]
-    year = date.year
-    return f"{day} {month} {year}"
 
 
 def get_time_string(date: datetime) -> str:
@@ -126,7 +117,7 @@ def get_unique_future_events(
     return unique_events
 
 
-def create_email_message() -> str:
+def create_event_email() -> str:
     text = requests.get(
         f"https://www.achurchnearyou.com/church/{CHURCH_ID}/service-and-events/feed/"
     ).text
@@ -134,9 +125,6 @@ def create_email_message() -> str:
     calendar = icalendar.Calendar.from_ical(text)
     events = [Event(eventData) for eventData in calendar.events]
     email = ""
-    if CHURCH_LOGO_URL:
-        email += f'<center><img style="width: 25rem;" src={CHURCH_LOGO_URL}></img>'
-        email += f"<h2>{CHURCH_NAME + ' ' if CHURCH_NAME else ''}Event Update - {get_date_string(datetime.now())}</h2></center>"
     email += "<h2>This Week's Events</h2>"
     email += f"<p>For full details of all events, please see our website on <a href=https://www.achurchnearyou.com/church/{CHURCH_ID}/>A Church Near You.</a>"
     this_weeks_events = get_this_weeks_events(events)
